Update patch set 21
Patch Set 21:

I've tested both this patch and one one in the master branch with just the resourceloader templates that are added in this patch.

Seems to have some caching problems with array_merge. Where as using it this way seems seems to not be such an issue.

I will go back to array_merge since this way won't be necessary, but improves some things like catching.

Patch-set: 21