Update patch set 5
Patch Set 5:

@Jdlrobson Well, in a perfect world, MobileFrontend would not exist, and core would work nicely with mobile devices out of the box. 

I also agree with avoiding hacks and cross-dependencies. I did not think that using pagerterms would be a problem, since it is already used in SearchGateway.js (at least in the beta version). 

I was just as surprised as you by this. I didn't even realize Special:Nearby was part of MF, I would have expected it to be part of GeoData. When I found out, I thought it would be easy enough to fix this, sicne MF is using wikidata/pageterms already anyway.

Of course, we (wikidata) should keep "our" quick hacks in "our" code base. But we didn't find an extension point to make this work for Special:Nearby. Your patch works in that direction, but it doesn't seem to be sufficient for integrating wikidata labels.

Patch-set: 5