Update patch set 1
Patch Set 1: Code-Review+1

Assuming this scenario is just testing that an anonymous user is taken to the login page on clicking a watchstar, then LGTM. Although I'm not really sure what the usefulness of checking for the presence of the message box is (is it just to verify you're on the login page? Isn't there a clearer way of testing that?).

Patch-set: 1
Reviewer: Jhobs <1948@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1