Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

@jdlrobson this looks a lot better, thanks!

But one distinction is missing: for wikidata.org itself, we not only want the descriptions, we also want to show the label instead of the actual page title or displaytitle. On Wikipedias and other client wikis, we want to use the descriptions, but not override the titles with wikidata labels.

See also the inline comments.

Patch-set: 4
Label: Code-Review=-1