Update patch set 16
Patch Set 16: Code-Review-1

(8 comments)

I went more in-depth in a comment on WatchstarGateway.js, but the purpose of this patch does not seem clear to me. Overall, it just appears to primarily rename "api" to "gateway," not actually change the behavior of anything. If that's the point, then great, but there are changes I think should be made to the gateway in the future.

Patch-set: 16
Label: Code-Review=-1