Update patch set 8
Patch Set 8:

there is no new branch until next Tuesday.  in an ideal world, it would be nice to have this now or yesterday ;) but ok... 

and big thanks for your help in working through these issues with us. Having a flexible solution that works also for places like the mobile watchlist is great.

Patch-set: 8