Update patch set 2
Patch Set 2:

In fact, we add 43 lines of new code, minus 12 lines of comments (which are removed from ResourceLoader) and the code optimisation and removal of whitespaces and so on (also done by ResourceLoader) we add approximately maybe 10-15 lines of new code, which, with compression, shouldn't be significantly more data to load in an async request (and that "only" one time, if it's properly cached). But, on the other hand, we have a complete DOM element and it's structure, which is added on _every_ page load. I'm not sure, if this alone will ever get a visible effect to the user, but if we, for every little thing, think, that this isn't visible to the user we sometimes have many little things which sum up to a huge and overloaded page DOM and maybe a performance killer (or at least a visible problem on the site, especially on small(er) devices which less power).

I added Sam and Rob to read their opinions, too. And I was free and added Krinkle, too, and hope, that he finds a little bit of time to give his opinion from a "performance team member" perspective :)

Patch-set: 2