Update patch set 1
Abandoned

I need to be able to focus some time on this or find someone to help me pair on making this happen - I can't do this alone.

The way I see this is happening is as follows:
* We set up a MinervaNeue extension on Gerrit
* We get that security reviewed and deployed.
* We test in production that the new skin works
* We remove the code from MobileFrontend

Anyone want to help me with the above? I will happily review as much as I can?

Patch-set: 1
Status: abandoned