Update patch set 24
Patch Set 24: Code-Review-1

Still not sure which files are not being tested and not sure what we gain by running this all on the Gruntfile which is not production code. Paladox I do ask that you take the time to raise bugs and talk about issues rather than submit patches. The status quo is fine with me and I personally dont think this change is useful.

Patch-set: 24
Label: Code-Review=-1