Update patch set 24
Patch Set 24:

(1 comment)

@Jdlrobson: There's definitely a few things here that were done too quickly because he didn't fully understand the context of each grunt subtask (like using jshint.all for unrelated qunit-coverage).

However, I don't see we wouldn't run JSHint and JSCS on the whole repo (excluding upstream libs of course). Why should JSCS and JSHint not pass on Gruntfile, it's just as much part of the code we maintain and don't want cognitive overload and inconsistency as anywhere else. It was actually my suggestion he do that when he asked me about it.

Virtually everybody else uses JSHint this way, both with and without Grunt, inside and outside Wikimedia. Whitelisting files is unusual and prone to error. There's no need to maintain a whitelist for these kind of things. It's the same thing you'd do on the command line or in an IDE. cd project; jshint .; The rest is handled by jshintrc and jshintignore.

Patch-set: 24