Update patch set 1
Patch Set 1:

Besides adding a config variable, is there anything else I need to do to test? What I did was to add the config variable and load the blacklisted page in a browser. I saw the desktop version, and when I clicked on the mobile view link at the bottom of the page, I saw the mobile version.

Patch-set: 1