Update patch set 1
Patch Set 1:

The headings of the sitelink groups (eg. Wikipedia/Wikisource etc. are <h3> elements which we don't want to appear in the toc. However, they contain a <span> which contains the number of sitelinks added to that group on the item. (It only gets filled by js which is another issue.)

So we have an element which should not appear in the toc. However, because it contains an unrelated <span> the parser tries to add it as a valid section. There seems to be no real option in here to exclude certain headings, besides them not containing a <span> element.

Patch-set: 1
Reviewer: Bene <549@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0