Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Couple questions and one small typo. Additionally, I know this is the "correct" way to do things, but this is a lot of overhead for 3 lines of CSS. We should figure out a breakpoint in the future of when the overhead is actually a worse performance hit, based on # of applied styles, % of pages that contain the affected element, etc. We could be actually hurting performance here (on top of requiring more code writing) by loading an additional script.

Patch-set: 2
Label: Code-Review=-1