Select headlines by class="mw-headline" instead of <span> element
This is much more specific, since it does not simply select *all*
<span> elements, no matter what role they have.

If there is more than one <span> in a headline, only the actual
headline element should be shown in the table of contents.

If there is *no* <span> marked with class="mw-headline", no other
<span> should be shown instead.

This fixes an actual issue with the HTML Wikibase produces on purpose.

Bug: T110902
Change-Id: I5c989f7657603a44b6ed87f0d364b434bb0d1b33