[SECURITY] Fix XSS vector when using fallback for displayTitle
The displaytitle pageprop is defined by core to be HTML,
and is treated by MF to be HTML. However, when no displaytitle
is set, the displaytitle in MF falls back to the plain page title,
or to pageterms.label, if available, both of which are plain text.
Thus, HTML escaping is necessary when applying the fallback.

Without this patch, <script> tags in a page title or wikibase label
will get executed on Special:Nearby, the mobile search suggester, etc.

Bug: T119707
Change-Id: I3a550a0e3a40813658462e49b24efe8145b28079