Update patch set 4
Patch Set 4:

Thanks for the patch Daniel. Since this is in gerrit, it's public to anyone who knows how to look for it-- so let's get this deployed as soon as we're reasonably sure it isn't going to break anything else.

Patch-set: 4
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0