Vary HTML output by NetSpeed designation
Vary MobileFrontend's HTML output based on connection speed, as indicated by
'NetSpeed' cookie. 'NetSpeed' may be set to either 'A' (for fast connections)
or 'B' (for slow connections). If cookie is not set or has an invalid value,
the default designation is 'A' (fast), simply because that is the current
implicit default.

When the NetSpeed designation is 'B', omit the `srcset` attribute from image
tags, and set image quality to 'low'. Currently this affects only the JPEG
encoder.

Bug: T119797
Change-Id: I8a1fb69724b2581c2efc1f5a9eddf72d4c10a6d2