Update patch set 24
Patch Set 24: Code-Review-2

(6 comments)

I remember, that I've wrote somewhere, that it would be better to make this in separate steps. But from what I've seen so far, the patches doesn't really prepare MobileFrontend for extension registration, mostly you just move code from a to be and vice versa and change a function here, a definition there, but all this doesn't (compared with the commit messages) convince me, that this is the right way. I would say, that the "last" change (without looking at the dependency patch here) is still a huge change without any minification. I don't hope that, but if so, please elaborate what and why you do in the last changes here using the commit messages (a real explanation, not just some words which, in fact, says "It works so"). Try to make things as clear as possible to help me (and others) to understand the deeper sense of the work you invest. Otherwise it seems to be more productive to have one change were the work is done to convert MF to extension registration, instead of having these (sorry for my words) fake changes :] (If you want, I would like to try, what problems we have and maybe upload a change which converts MobileFrontend to extension registration).

Patch-set: 24
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2