Update patch set 4
Patch Set 4: Code-Review+1

There is a minor UI consistency for users who disabled images before this patch - visiting the special mobile options page shows images as enabled. Unticking box keeps images disabled but on second visit the page is as expected (so essentially to enable images again they have to visit the page twice). I think this is acceptable and happy to merge despite this edge case but we should state that is the case in the commit message.

Patch-set: 4
Label: Code-Review=+1