Update patch set 24
Patch Set 24:

In general I wonder if this is worthwhile and whether all this work is causing us more work on the long term. I know I've investing a lot of time and energy into these patchsets.

We would like to move Minerva out of MobileFrontend and use skin registration from the offset and this complicates that (I'm hoping early 2016).

Paladox would be great if you could take a pause and reflect on whether these changes are helping make this happen any faster.

Patch-set: 24