Update patch set 1
Patch Set 1:

(1 comment)

In the arrow-down.svg there is some space above the arrow but not below it. When transforming the svg to be an up arrow the arrow shifts up because of the lack of space above it. Could you also change the svg and put the arrow in the middle of the view box?

Patch-set: 1
Label: Code-Review=0