Update patch set 3
Patch Set 3:

The feature is the same.. it shows a preview of content before you click it, thus it should be in Popups. Arguably it could be a separate feature but would such a thing be useful on another skin such as Vector? If so what is it doing differently?

> It would create a cross dependency to MobileFrontend. It would be possible, but would need additional tests in Popups to make sure we don't get the same desaster as with the Thanks extension.

True, but this is easily done by adding MobileFrontend to the build which runs unit tests/browser tests etc...
I know it's a pain but it's the right thing to do if we are going to pull apart MobileFrontend.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0