Update patch set 1
Patch Set 1: Code-Review-1

The solution may solve the problem for a particular device, but it certainly won't work on Asus Nexus 7, which has the device width of 601px and device-pixel-ratio of 1.331.

Rather than changing 768px we should explore taking device-pixel-ratio into account. In Galaxy S5 mini case, looks like the ratio is 2, which effectively makes the device width 1440px.

Patch-set: 1