Update patch set 1
Patch Set 1: Code-Review-1

Im not a 100% sure if this will show just to autoconfirmed users - on my local wiki it shows to everyone even anons which was not the intention of this task. We need a custom mobiletalk permission e.g. ->user->isAllowed( mobile-talk ) that defaults to autoconfirmed in case we need to rethink this (for example due to increased talk page spam

Patch-set: 1
Label: Code-Review=-1