Add error handler for Nearby page search
If a page does not exist, Nearby currently simply hold up in the "Loading"
status, which is, in fact, false and misleading.

This change adds a new handler for the NearbyGateway#_search function to reject
the returned Deferred, when the api call is rejected. Aditionally, the error messages
taken in this case are, if provided, taken from the api response. If no info provided,
the 'server' (now known as 'http') error message is used.

Change-Id: Ia645fe7627150058b6007ca7e6252e44a416ee13