Update patch set 4
Patch Set 4:

> Doesn't this mean showing the first notification will be much slower?

In fact: yes. But it depends on: The module is 7,4 kB big, and takes 1,86s to load with GPRS (50 kB/s, with all the lookup and 500 RTT). This isn't as much as it needs to be, that the user is interrupted in the workflow, I think. Futhermore, e.g. for the watch action, the indication, that the page is watched now takes a long time, too, so the user normally should expect, that any other visually indication could take a bit longer. :)

On the other hand: This change reduces the overall size of the _first_ request of a page by 7,4kB (less or more), which is a good fact, too, because a notification toast isn't needed at any page load, right? :)

Patch-set: 4
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0