Add Vary header User-Agent, if appropriate
According to Google's best practices [1], MobileFrontend should add the
Vary header User-Agent, if it is configured to autodetect the mobile view
and use the same url for desktop and mobile.

The feature is feature-flagged and need to be enabled with the new configuration
variable $wgMFVaryOnUA (the other conidition must still match).

[1] https://developers.google.com/webmasters/mobile-sites/mobile-seo/dynamic-serving?hl=en#the-vary-http-header

Bug: T123189
Change-Id: I8165994dea379eee93500e4aada48234679e7513