Update patch set 2
Patch Set 2:

The bug this tries to solve seems to be a Chrome only issue and not reproable on Firefox and Safari. The issue is the leaved contentDir on "content.hogan" template and this tries to solve the issue by filling them like the way happened on skins.minerva.editor/init.js#L140.

Patch-set: 2