Update patch set 1
Patch Set 1:

Just a quick look: The better question is, why RequestContext::canUseWikiPage() returns true. This is used to make sure, that RequestContext::getWikiPage() can be used without throwing an error and is correctly called before using it[1]. I think the change makes sense and I will merge it, after I had the chance to test it. However, it would be nice to track down my note, too, to make sure, this isn't an error which might come back later :)

Btw.: RequestContext::canUseWikiPage() used $title->canExists() as far as I know (haven't had the time to check, yet). And if ApiMain sets the title to a special page, the Title::canExists() function should return false.

[1] https://github.com/wikimedia/mediawiki-extensions-MobileFrontend/blob/master/includes/MobileFrontend.body.php#L47

Patch-set: 1