Update patch set 1
Patch Set 1:

Ah, wait, it seems, that DerivativeContext uses the title, which is set by the function's hack, and only if it wasn't set, uses the RequestContext title[1]. The DerivativeContext::getWikiPage on the other hand, uses the RequestContext instance directly without using the title set to the DerivativeContext[2]. In my point of view, this should be a bug in mediawiki/core, too, because of O set a title to a derivative context, I would assume, that it's used for any function, that involves the title, too. Like getWikiPage().

[1] https://github.com/wikimedia/mediawiki/blob/master/includes/context/DerivativeContext.php#L188-L196
[2] https://github.com/wikimedia/mediawiki/blob/master/includes/context/DerivativeContext.php#L217-L223

Patch-set: 1
Label: Verified=0