Make SpecialMobilePreferences more generic for non-core sections
Currently, Special:MobilePreferences shows and renders core-preferences,
only, which is fine for the selection list on the entry point. However,
this prevents other extnesions to link to a specific preferences section
they add to the preferences special page in mobile mode.

This change tries to make the preferences handling a bit more generic and
doesn't rely on any hardcoded list of valid sections anymore (for rendering
only). This allows other extensions to link to a specific section inside
Special:Preferences with the section name (which is used in the section
key of the HTMLForm form descriptor).

Change-Id: I6d8c5b80e70c3d8d1a92a70cc91e1b90d598cb0f