Hygiene: Remove legacy HTML removable class option
This seems to have been previously used for treating API queries differently
from page HTML. I don't see a clear use case for this and this can lead to
confusion, so let's not treat them any different.

I suspect it is no longer needed anyway
given there is nothing in mediawiki-config configuring it.

Change-Id: I19221697114586aa78235ae688cb7b39cb526949