Update patch set 8
Patch Set 8: Code-Review-1

Hmm, instead of having this in MobileFrontend, I would love to see this in mediawiki/core. This isn't useful for mobile devices, only. Someone may want to have this for a netbook, too, or for the notebook when he's online with a limited data connection. Having yet another special thing in mobile formatter only makes me unhappy :)

Patch-set: 8