Update patch set 8
Patch Set 8:

Sure thing Florian, there's been a lot of chatting with several members of varied teams (ori, bd808, etc) and we agreed to implement this for mobile first and measure benefits, but with looking forward to upstreaming. It's easier to measure the benefits for the bulk of people using the 2G & 3G networks on the mobile site and see how we have to tweak it. Initially, we're not even going to deploy this anywhere, that's why we're feature flagging it. We're going to put it on the beta cluster and start measuring there.

That's why I have to move the image transformation html code to parser output hooks and not MobileFormatter (TODO, see commit message), and I have to extract the lazy loading images JS to it's own module so that we can extract it and upstream it more easily (not yet on TODO, I'll add it now!).

More info in the phab bug and blocked tasks/goals. This will be for desktop and mobile, we're not sure in what form yet though, since we have to implement, measure in real settings, and decide how to roll this out to stable both in desktop and mobile.

Thanks for the comment!

Patch-set: 8
Reviewer: Jhernandez <2027@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0