Update patch set 2
Patch Set 2: Code-Review-1

Im wary of putting this in the util.js which is lin the initial page load without a clear idea of the motivation and use cases. If its specific for the language overlay work place it in LanguageOverlay which only loads when you click on the button. If you see more use cases please open a task so we can discuss this some more.

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1