Don't limit mw-ui-input border on Special:UserLogin
Not all fields (especially fields which are generated by extensions) have
has a mw-ui-vform-field class in the div-container. But these fields should
have a border, too, so an user can identify them as an input field.

This change removes a redundant border: rule and moves the border rule for
the mw-ui-vform-field input fields to apply to all mw-ui-input fields.

Bug: T125945
Change-Id: I3261dfcd5bb289706e10cb3cd758ca79aaaedc5c