Update patch set 3
Patch Set 3:

I think we should make it an experiment and allow lazy loading images so for some clients and not others to carry out an A/B test if we wish. That said, the patch looks good.

Patch-set: 3
Reviewer: Bmansurov <1963@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0