Update patch set 4
Patch Set 4:

If you call "for the majority of cases" main namespace pages, only, ok, but I don't think, that this is correct. In fact, the bug described in the task is "only" one (and maybe the more important) incarnation of the underlying problem, but in general, any code that relies on the lead section (and that it is wrapped into a div) is affected (in any way). Rather than fixing the problem in a "frontend" code (the edit function is an user of the leadSection handling function/api of MobileFrontends backend, if you want to see it in this way) I would prefer to fix the underlying problem (even if I don't be a fan of touching MobileFormatter, too).

Patch-set: 4