Provide low-resolution NetSpeed option in Special:MobileOptions
After chatting with Ori,
given the images saving in bytes, the lower fully loaded time
(35s to 22s) and the fact the code is written let's make
this an available option to our users.

On long term it's probably a better option than disabling images altogether
(We may want to remove the image disabling feature if we see a drop in usage
as a result of this new option)
Have okayed with Jon Katz and Nirzar.

Change-Id: I1fe8538eb679fb7397f482407ba88c4286174e39