Update patch set 3
Patch Set 3:

I'm sorry, I'm not sure as to the broader context, and perhaps there are other things we can do here to accomplish whatever it is, but the bottom line is we definitely don't want WMF wikis emitting "Vary: User-Agent" towards our varnishes.  I'll try to make time to look at the ticket and the google link in more detail later.

Patch-set: 3