Update patch set 18
Patch Set 18:

I've created https://phabricator.wikimedia.org/T126590 to block putting this patch to production, and made a few tasks with the unknowns/questions from this version so that we don't forget to address them before deploying to prod.

The current version seems good enough to run tests on beta cluster. Thoughts?

Sub tasks:

* https://phabricator.wikimedia.org/T126593
* https://phabricator.wikimedia.org/T126592
* https://phabricator.wikimedia.org/T126591

I'm blocked if we have to address the above problems in this patch, and would really appreciate help. I'm also leaving on a work trip for 2 weeks so I'll need somebody to take ownership of this from next week onwards.

Patch-set: 18
Label: Verified=0