Allow a user to change the skin with the api
With the mobileskin option, the user can specify, which skin he/she want to use.
The option if available through the api and can, e.g., changed using this query
in the brwosers console (while the tab with the target wiki is open):

new mw.Api().postWithToken(
	'options',
	{
		action: 'options',
		change: 'mobileskin=vector'
	})
	.fail( function( error ) {
		console.log( error );
	})
	.done( function( resp ) {
		console.log( resp, 'done' );
	});

After a short waiting time, the console should show a "done", if anything worked
correctly. After a reload, the mobile site should use vector as the skin class, instead of minerva.

Change-Id: I14bf3f80ad5d935a6ea4ed61f85dd9e9b02b433d