Update patch set 9
Patch Set 9: Code-Review-1

See http://imgur.com/a/AWkqF I am fine with this in principle but I worry that this is not the right place for this action. Can I maybe suggest getting some feedback from a designer - might this be better for example after the next and previous diff links? ps. i removed Resources.php :)

Patch-set: 9
Label: Code-Review=-1