Allow lazy loading of references (feature flagged)
When enabled this replaces the list of references with
a link to a non-existent special page which is passed
the parameters it needs to render the list of referenes

When a reference is clicked it fires off a request to the API
to get the references object.

This needs various changes in the Cite API which are listed
below.

Changes:
* When feature flag enabled sub and sup tags are hidden. URLS for
these will be rewritten when Special:References fallback is in place
* setup function of references module  now requires the page object
* getReference private method now requires page to be passed
* Introduce private method getReferenceData which obtains
references via API or from DOM depending on feature flag
* mobile.references now uses the editor gateway so it can parse
reference raw wikitext (hopefully this can be removed at later date)
* Refactor tests for reference to use a template and not trigger
DOM events
* New style module added for beta skin
* Error handling when reference lookup fails

Depends-On: I81a965bcb47d17df18f1e415e3c25f88f6b48ffc
Bug: T125896
Change-Id: I287608a8fdfa61fe2d9270ab7df4da9bca3aef4c