Update patch set 3
Patch Set 3:

> we definitely don't want WMF wikis emitting "Vary: User-Agent" towards our varnishes

Reasonable, but as pointed out: This doesn't apply to WMF wikis as these wikis use different urls for mobile and desktop version of a page. This applies to pages with the _same_ url for both, desktop and mobile pages, only.

I would welcome your feedback about the google link, maybe I misinterpreted it, but I understand it so, that you should send the vary header with user-agent, if you use the same url for both content versions :)

Patch-set: 3