Don't call isBetaGroupMember in ResourceLoaderGetConfigVars hook
Remove code relating to beta, given that the default behaviour is to
enable this now. This should have been cleaned up when pushing to stable

Add a note that isBetaGroupMember should not be called in this function
given its result needs the User object to work.

This aside MobileContext is perfectly safe in this environment

Bug: T127860
Change-Id: I022d41a6fbf2ee2da32939656a77e85ce26e063b