Don't use decodeURIComponent on Undefined
decodeURIComponent will return the string "undefined", which casts to
the boolean true. That means, that the talk overlay would add new discussions
to the title "Undefined" instead of the expected mw.config.get( 'wgPageName' ).

Fix this by check first, if the title data attribute is Undefined (which casts to
the boolean false) and if not use decodeURIComponent.

Bug: T128273
Change-Id: I59dbbe02f25bf123ebe3cae947289c611a23b9dd