Update patch set 1
Patch Set 1: Code-Review-1

These should work as they work locally.
If we remove this we will have no coverage whatsoever unless someone manuall runs them.
The team currently consists of 2-3 people so we are a little stretched to fix these right now but we should fix them soon, otherwise there is no point in them being in the repository.
The failing tests at least should serve as a reminder to fix this.

Patch-set: 1