Update patch set 1
Patch Set 1:

I don't know how often this code path is run, but if it's close to every page view from logged-in and/or some logged-out users, you should probably sample it.

Or, since this is instrumenting just a single method (is it okay to include the parent:: call?) this information might be already available in a better way via the sampled xenon or xhprof logs.

Patch-set: 1