Lazy loaded images in beta
Arguably we can do this now. This will not work for Grade C browsers,
but I suspect there are a low number of grade C browsers in mobile.
Also beta users should expect experimental software.

We should wait on outcome of T126793 and preferably merge after
but lets not block merging this sprint... the bugs we can uncover by
putting this in beta are far more important.

(Note not done in beta given this is expected to be a default recommended
config in future)

Bug: T126590
Change-Id: I80dd9d92d42a674c57e98800277caa0e0d30b59e