Perform sectioning in MobileFormater via DOM
So that transformations can be given information about which section they are
in so that they can decide if they should apply or not.

This will be used for applying or not the lazy loading images transformation
for example only after the lead section in a follow up commit.

DOM implementation based off @phuedx's fork:
https://github.com/phuedx/mobile-formatter

* Breaking change:
  * Section bodies are now identified with a class like "mf-section-4" which
  shows the number of section it is (0 based).
  * "mw-mobilefrontend-leadsection" removed in favor of "mf-section-0"
 but checked in Page.js for backwards compatibility

Bug: T127128
Change-Id: I0675be5dcaab71e7693090e5c7114fe5c8339586