Add Grade C browser support to lazy loaded images
Placeholders are instantly shown to Grade C browsers to avoid jumpy content
but are hidden to real no-javascript browsers.

A script tag is added to the bottom of the page to load all images for
Grade C browsers using the NORLQ queue, which is only run by Grade C
browsers with JavaScript enabled.

Bug: T126793
Change-Id: I5ca4b4eb63bdf02743b8ba95c59279735c240e46