Update patch set 7
Patch Set 7:

This is not compatible with either of the messages in the WikimediaMessage extension :/ Im not sure what to suggest to get this working. We may need to make some changes in core or rethink the message in MobileFrontend to make it more compatible with the default. Any ideas?

Patch-set: 7