Update patch set 1
Patch Set 1: Code-Review+1

Oh dear, good catch.

If we were to configure $wgIncludeLegacyJavaScript and $wgPreloadJavaScriptMwUtil to false on the mobile site it seems that this would avoid the load of the 4.9kb and the additional request.

Even though this is being fixed soon - May is still 2 months away... would that not be a better solution, given these are not needed on mobile?

Patch-set: 1
Label: Code-Review=+1