Lazy-load: Avoid jQuery filter() due to pushStack() memory
Assigning the result of filter() is the same as chaining.

For chaining support, jQuery has to keep a stack of all previous versions of
the collections so that things like end() and addBack() work as expected.

In general, when maintaining a collection for a long time, avoid jQuery methods
and use a plain array instead. Otherwise it requires jQuery to keep hold of
all previous versions of the array.

Change-Id: I6b0f08f3f32b490110c7faf5c649effdb3818936