Update patch set 8
Patch Set 8:

(2 comments)

Should be fine. A few minor points. As for the gradec class. That shouldn't be needed since the elements are swapped unconditionally at load here. So whatever the class is used for, the placeholders won't be visible for long in that state.

I say "not for long" because they will flash briefly. I suspect you want the class to avoid that, however that won't work because the core patch applies the class from startup, which runs async and probably at the same time as this code (note how the core patch is near the code that runs NORLQ, which this patch uses).

There is no way currently to change the layout ahead of time because this depends on the startup module, which we've made async. It's a minor FOUC (if visible above the fold) but mostly unavoidable. This is imho acceptable as Grade C experience and restores itself generally with a second.

Patch-set: 8
Label: Verified=0