Specify Special:EditWatchlist in watchlist switching tests
Switching to List view will fail if run after
Switching to Feed view as a visit to Special:Watchlist
will redirect to the last page visited in Watchlist and break
the test.

In Sauce labs tests seem to be hanging on the Watchlist step.
e.g.
https://integration.wikimedia.org/ci/view/Mobile/job/browsertests-MobileFrontend-en.m.wikipedia.beta.wmflabs.org-linux-chrome-sauce/985/testReport/junit/(root)/Manage%20Watchlist/Switching_to_Feed_view/

This step seems to be navigating to Special:Watchlist via
a click to the hamburger menu instead of a URL change.

It's not sure if this should fix the bug tagged but it's more
correct and may.

Bug: T94156
Change-Id: I6aec7c68c3c384383fa9751b8a94a879a90a2897