Update patch set 17
Patch Set 17:

Verified on Chrome with MSIE7 as user-agent string, and verified in BrowserStack with real IE7 and real IE8.

Remove the Array#slice() for now since that fails in IE7/IE8. We can guard it with try/catch and fallback to direct NodeList iteration, but that seems a bit pointless as it might as well be left out at that point.

Patch-set: 17