Update patch set 2
Patch Set 2: Code-Review-1

Github is not the canonical repository for the MobileFrontend but merely a mirror.  Additionally, I dont think it makes much sense to point to a VCS system in package.json since that extension is unlikely to be uploaded on npmjs.com as a package.

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1