Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I've verified this locally and it LGTM.

Sorry for not addressing this in my previous review but could you include an identifying prefix in the cookie name please, e.g. mf_LazyLoadImages? It makes it quite clear that this cookie affects the behaviour of the MobileFrontend extension and follows the convention for the "useformat" cookie.

 There's the inline comment from Krinkle, which, generally, I agree.

Patch-set: 3
Label: Code-Review=-1