Update patch set 8
Patch Set 8: Code-Review-1

Even on Firefox on Android (width > 320px) the title shifts to the left a little bit. We should explore another way of doing this. For example, even though we can fix this piece of FOUC the section collapsing happens on small screens which results in shifting of text. Rather than collapsing section with JS we can use CSS to collapse sections for JS supported browsers and small screens, and not collapse when the screen is tablet-size or JS is not supported. I hope these changes will make this particular problem easy to fix too.

Patch-set: 8
Label: Code-Review=-1