Update patch set 8
Patch Set 8: Code-Review-1

We should take this opportunity to fix the TOC once and for all. Why don't we support the TOC on non JS browsers for example? Why can't we progressively enhance it? Why not generate the TOC in the backend and output it for all pages, while styling it differently (collapsed vs expanded) for JS / Non-JS browsers or different form factors? This change feels like another hack to fix a problem at hand and not fix the bigger problem which we have.

Patch-set: 8
Label: Code-Review=-1