Update patch set 9
Patch Set 9:

I'm simply doing what adam suggested here: https://phabricator.wikimedia.org/T126836#2124052

> We should take this opportunity to fix the TOC once and for all. Why don't we support the TOC on non JS browsers for example? Why can't we progressively enhance it? Why not generate the TOC in the backend and output it for all pages, while styling it differently (collapsed vs expanded) for JS / Non-JS browsers or different form factors?
We made the decision a while back that collapsed sections made the need for a table of contents redundant on mobile. It also bloated the HTML size, in particular for large articles.

> This change feels like another hack to fix a problem at hand and not fix the bigger problem which we have.
This is thus not a hack, it was an intentional change, although poorly thought out.

I'd love to but a rewrite of how we do table of contents will require major design input, data evaluation and user testing.

This is a big discussion about the future of table of contents and it will be fuelled further by the analysis Tilman is doing right now (https://commons.wikimedia.org/wiki/File:New_readership_data_(Wikimedia_Foundation_Tech_Talk).pdf). This is unlikely to change anytime soon though so we should do a better job of the existing code given the constraints we have.

Can you articulate what is wrong from a technical point of view with this solution and how you would improve it.
Have you time to re-evaluate bloating the HTML by containing the entire table of contents for 2G users? Do you have time to untangle the ResourceLoader module that comes with it (https://phabricator.wikimedia.org/T130632)?

Patch-set: 9
Label: Verified=0